requirements.txt.txt
Flask
speechrecognition
gtts
requests
Pillow
sumy
transformers
pvporcupine
pyaudio
torch
torchvision
